+++ CHANGELOG.md
@@ -15,9 +15,9 @@ Increment the:
 
 ## [Unreleased]
 
-* [SDK] Fix lost-wakeups in BatchSpanProcessor to prevent stalls during
-  shutdown and force flush.
-  [#4382](https://github.com/open-telemetry/opentelemetry-cpp/pull/4382)
+* [CONFIGURATION] Apply metric view `attribute_keys` include and exclude patterns
+  when building the SDK, with exclude patterns taking precedence.
+  [#3546](https://github.com/open-telemetry/opentelemetry-cpp/issues/3546)
 
 * [CONFIGURATION] Add support for the composite sampler configuration
   (programmatic and from yaml)
@@ -172,6 +172,10 @@ Increment the:
 
 * [CONFIGURATION] file configuration - yaml schema 1.1.0
   [#4340](https://github.com/open-telemetry/opentelemetry-cpp/pull/4340)
+
+* [SDK] Fix lost-wakeups in BatchSpanProcessor to prevent stalls during
+  shutdown and force flush.
+  [#4382](https://github.com/open-telemetry/opentelemetry-cpp/pull/4382)
 
 Breaking changes: